+++ corehq/apps/translations/templates/app_translations.html
